The PJIT now takes a parameter where it writes its result to, rather than returning a result.
We pass a stack pointer to this parameter. The PJIT then writes the optimized version of the function into the stack address once it's ready (immediately if it was already ready at time of call). The caller can in theory poll its stack variable to see if the optimized version is ready.